pyproject.toml
[tool.poetry]
name = "hiro"
version = "0.1.0"
description = ""
authors = ["Siarencjusz"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.8"
numpy = "^1.23.1"
pylint = "^2.14.4"
jupyter = "^1.0.0"
jax = "^0.3.14"
optax = "^0.1.3"
dm-haiku = "^0.0.7"
wandb = "^0.12.21"
pymongo = "^4.1.1"
pymongoarrow = "^0.4.0"
matplotlib = "^3.5.2"
Pillow = "^9.2.0"
dm-pix = "^0.3.2"
tqdm = "^4.64.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
